fix: display legacy sections separately
