+++ src/components/Administrator.js
@@ -1,9 +1,9 @@
 import React from "react";
 
 export const Administrator = () => (
-  <div class="mb-5">
-    <span class="border-rose-800 border-2 bg-rose-100 border-solid text-rose-700 py-1 px-2 rounded-md">
-    Role Administrator
+  <div className="mb-5">
+    <span className="border-rose-800 border-2 bg-rose-100 border-solid text-rose-700 py-1 px-2 rounded-md">
+      Role Administrator
     </span>
   </div>
 );
