+++ src/components/Recommended.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 export const Recommended = () => (
-  <span class="bg-sky-700 px-2 py-1 rounded-md text-white text-xs">
+  <span className="bg-sky-700 px-2 py-1 rounded-md text-white text-xs">
     RECOMMENDED
   </span>
 );
