+++ src/pages/docs/configuration.js
@@ -5,31 +5,22 @@ import { Link } from "../../components/Link";
 import { Main } from "../../components/Main";
 import { Nav } from "../../components/Nav";
 
-const Legacy = ({ data }) => {
+const Manual = ({ data }) => {
   return (
     <Main>
       <Nav />
       <div className="lg:container mx-auto gap-2 grid grid-cols-9">
-        <div className="col-span-1">
+        <div className="col-span-2">
           {data.allMdx.nodes.map((node) => (
-            <div>
-              <Link
-                key={node.frontmatter.title}
-                to={`/docs/legacy#${node.slug.replace("legacy/", "")}`}
-              >
-                {node.frontmatter.title}
-              </Link>
-            </div>
+            <span>
+              <Link to={`/docs/${node.slug}`}>{node.frontmatter.title}</Link>
+            </span>
           ))}
         </div>
-        <div className="col-span-6 flex flex-col bg-neutral-50 font-sans relative ">
+        <div className="col-span-7 flex flex-col bg-neutral-50 font-sans relative ">
           {data.allMdx.nodes.map((node) => (
             <div>
-              <h1 className="text-blue-700 text-5xl mb-10 mt-5">
-                <a href={`#${node.slug.replace("legacy/", "")}`}>
-                  {node.frontmatter.title}
-                </a>
-              </h1>
+              <h1>{node.frontmatter.title}</h1>
               <MDXRenderer>{node.body}</MDXRenderer>
             </div>
           ))}
@@ -40,8 +31,8 @@ const Legacy = ({ data }) => {
 };
 
 export const query = graphql`
-  query MdxQuery {
-    allMdx(filter: { fileAbsolutePath: { regex: "/legacy/" } }) {
+  query manualQuery {
+    allMdx(filter: { fileAbsolutePath: { regex: "/manual/" } }) {
       nodes {
         body
         slug
@@ -54,4 +45,4 @@ export const query = graphql`
   }
 `;
 
-export default Legacy;
+export default Manual;
