+++ src/pages/docs/index.js
@@ -31,7 +31,7 @@ const DocsIndex = () => (
           </div>
           <div>
             <h3>
-              <Link to="/docs/legacy">Legacy API Reference</Link>
+              <Link to="/docs/manual">Legacy API Reference</Link>
             </h3>
             <p>API Reference for legacy Virtool releases prior to 5.0.0.</p>
           </div>
